♻️ refactor(server): converge webhook / dev / composio routes into hono
Follow-up to the agent-eval-run move: the remaining plain-handler route
segments under `src/app/(backend)/api` become hono apps under
`apps/server/src/router-hono`, served through per-prefix catch-alls. All URLs
are unchanged.

- `webhooks/*` (casdoor, logto, memory-extraction ×2, memory-user-memory ×2,
  video/:provider) → `router-hono/webhooks`, one `api/webhooks/[[...route]]`
- `dev/*` → `router-hono/devtools` behind `api/dev/[[...route]]`. Named
  `devtools` because `router-hono/dev.ts` is the standalone dev-server
  bootstrap and would shadow a `dev/` directory in path resolution.
- `composio/oauth/callback` → `router-hono/composio`

`api/auth/**` is deliberately left alone — Better Auth and the two pre-auth
lookups that sit beside it stay Next.js route segments.

Two guards that were copy-pasted per route become middleware:
`memoryWebhookAuth` (the configured static-header check, shared by 4 handlers)
and `devOnly`.

Route segments under `api/`: 29 → 11.

Since the directory layout no longer *is* the URL, a path typo now 404s at
runtime instead of failing the build. Adds mount-table tests pinning the
webhooks and devtools apps; the moved webhook tests drive their handler through
a one-route hono app rather than the full index, which would pull in
model-runtime and the db clients.
