+++ apps/server/src/router-hono/composio/index.ts
@@ -1,0 +1,9 @@
+import { Hono } from 'hono';
+
+import { composioOAuthCallback } from './handlers/oauthCallback';
+
+const app = new Hono().basePath('/api/composio');
+
+app.get('/oauth/callback', composioOAuthCallback);
+
+export default app;
