+++ src/app/(backend)/api/composio/[[...route]]/route.ts
@@ -1,0 +1,3 @@
+import app from '@/server/router-hono/composio';
+
+export const GET = (request: Request) => app.fetch(request);
