+++ src/app/(backend)/api/dev/[[...route]]/route.ts
@@ -1,0 +1,6 @@
+import app from '@/server/router-hono/devtools';
+
+const handler = (request: Request) => app.fetch(request);
+
+export const GET = handler;
+export const POST = handler;
