+++ src/app/(backend)/api/webhooks/[[...route]]/route.ts
@@ -1,0 +1,3 @@
+import app from '@/server/router-hono/webhooks';
+
+export const POST = (request: Request) => app.fetch(request);
